### 2022-10-07 #4826 @editorialbot
[PRE REVIEW]: BellDiagonalQudits: A package for entanglement analyses of mixed maximally entangled qudits
**Submitting author:** @kungfugo (<a href="http://orcid.org/0000-0002-2352-3731">Christopher Popp</a>)
**Repository:** https://github.com/kungfugo/BellDiagonalQudits.jl
**Branch with paper.md** (empty if default branch):
**Version:** v.0.1.1
