### 2022-10-27 #4826 @jarvist
Glad to have you on board!
The Reviewer Guidelines are a good starting place: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
(And generally, the JOSS readthedocs site is where I go for authoritative answers!)
