### 2022-11-02 #4826 @kungfugo
@jarvist , @meandmytram  thank you for your effort!
Please note that my package is related to  _classical computing_ for _quantum systems_, not _quantum computing_ (as in executed on a quantum computer). In principal, Julia knowledge and some physical background should be sufficient to understand and review the package.
Of course, I am always available to answer any questions.
