### 2022-11-11 #4924 @editorialbot
[REVIEW]: BellDiagonalQudits: A package for entanglement analyses of mixed maximally entangled qudits
**Submitting author:** @kungfugo (<a href="http://orcid.org/0000-0002-2352-3731">Christopher Popp</a>)
**Repository:** https://github.com/kungfugo/BellDiagonalQudits.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.5
