+++ paper/paper.bib
@@ -5,7 +5,8 @@
   year={2022},
   volume={12},
   pages={12472},
-  URL={https://doi.org/10.1038/s41598-022-16225-z}
+  URL={https://doi.org/10.1038/s41598-022-16225-z},
+  DOI = {10.1038/s41598-022-16225-z}
 }
 
 @article{baumgartner,
@@ -15,7 +16,7 @@
     volume = 40,
     pages = 7919,
     year = "2007",
-    DOI = "https://doi.org/10.1088/1751-8113/40/28/S03",
+    doi = {10.1088/1751-8113/40/28/S03},
     url= {https://doi.org/10.1088/1751-8113/40/28/S03}
 }
 
@@ -59,5 +60,6 @@
   author={Beatrix C. Hiesmayr},
   journal={Scientific Reports},
   year={2021},
-  volume={11}
+  volume={11},
+  doi = {10.1038/s41598-021-98523-6}
 }
