### 2022-11-15 #4924 @kungfugo
@jarvist Thanks for pointing out. .bib now contains the correct DOIs.
