### 2022-12-02 57e84cf
Add installation instructions README. Add CI.yml
