+++ .github/workflows/CI.yml
@@ -12,7 +12,8 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        julia-version: ['1.0', '1', 'nightly']
+        julia-version: 
+          - '1.8'
         os: 
           - ubuntu-latest
 
