+++ src/Symmetries/symmetries.jl
@@ -1,11 +1,11 @@
 """
-    getPermutationFromTranslation(translation::Tuple{Int,Int}, stdBasis::StandardBasis, d)
+    get_permutation_from_translation(translation::Tuple{Int,Int}, stdBasis::StandardBasis, d)
 
 Return `Permutations.Permutation` of translation for ``d^2`` dimensional vector of coordinates in Bell basis `standardBasis`.
 """
-function getPermutationFromTranslation(translation::Tuple{Int,Int}, stdBasis::StandardBasis, d)
+function get_permutation_from_translation(translation::Tuple{Int,Int}, stdBasis::StandardBasis, d)
 
-    myDicts = createDictionaryFromBasis(stdBasis)
+    myDicts = create_dictionary_from_basis(stdBasis)
     dict = myDicts[1] #keys are tuples 
     reverseDict = myDicts[2] #keys are indices
 
@@ -29,14 +29,14 @@ function getPermutationFromTranslation(translation::Tuple{Int,Int}, stdBasis::St
 end
 
 """
-    getPermutationFromMomentumInversion(stdBasis::StandardBasis, d)
+    get_permutation_from_momentuminversion(stdBasis::StandardBasis, d)
 
 Return `Permutations.Permutation` of momentum inversion for ``d^2`` dimensional vector of coordinates in Bell basis `standardBasis`.
 """
-function getPermutationFromMomentumInversion(stdBasis::StandardBasis, d)
+function get_permutation_from_momentuminversion(stdBasis::StandardBasis, d)
 
     #load dictionary
-    myDicts = createDictionaryFromBasis(stdBasis)
+    myDicts = create_dictionary_from_basis(stdBasis)
     dict = myDicts[1] #keys are tuples 
     reverseDict = myDicts[2] #keys are indices
 
@@ -59,14 +59,14 @@ function getPermutationFromMomentumInversion(stdBasis::StandardBasis, d)
 end
 
 """
-    getPermutationFromQuarterRotation(stdBasis::StandardBasis, d)
+    get_permutation_from_quarterrotation(stdBasis::StandardBasis, d)
 
 Return `Permutations.Permutation` of qurater rotation for ``d^2`` dimensional vector of coordinates in Bell basis `standardBasis`.
 """
-function getPermutationFromQuarterRotation(stdBasis::StandardBasis, d)
+function get_permutation_from_quarterrotation(stdBasis::StandardBasis, d)
 
     #load dictionary
-    myDicts = createDictionaryFromBasis(stdBasis)
+    myDicts = create_dictionary_from_basis(stdBasis)
     dict = myDicts[1] #keys are tuples 
     reverseDict = myDicts[2] #keys are indices
 
@@ -89,14 +89,14 @@ function getPermutationFromQuarterRotation(stdBasis::StandardBasis, d)
 end
 
 """
-    getPermutationFromVerticalShear(stdBasis::StandardBasis, d)
+    get_permutation_from_verticalshear(stdBasis::StandardBasis, d)
 
 Return `Permutations.Permutation`` of vertical shear for ``d^2`` dimensional vector of coordinates in Bell basis`standardBasis`.
 """
-function getPermutationFromVerticalShear(stdBasis::StandardBasis, d)
+function get_permutation_from_verticalshear(stdBasis::StandardBasis, d)
 
     #load dictionary
-    myDicts = createDictionaryFromBasis(stdBasis)
+    myDicts = create_dictionary_from_basis(stdBasis)
     dict = myDicts[1] #keys are tuples 
     reverseDict = myDicts[2] #keys are indices
 
@@ -119,24 +119,24 @@ function getPermutationFromVerticalShear(stdBasis::StandardBasis, d)
 end
 
 """
[…]
