+++ src/SeparableStates/separableStates.jl
@@ -1,22 +1,22 @@
 """
-    createKernelVertexStates(d, standardBasis::StandardBasis)
+    create_kernel_vertexstates(d, standardBasis::StandardBasis)
 
 Return array containing collections of corresponding `standardBasis` indices, coordinates and density matrices for all `d` element sublattices in discrete phase space.
 """
-function createKernelVertexStates(d, standardBasis::StandardBasis)
+function create_kernel_vertexstates(d, standardBasis::StandardBasis)
 
-    allSubLattices = createDimElementSubLattices(d)
-    vertexStates = map(x -> createIndexSubLatticeState(standardBasis, x), allSubLattices)
+    allSubLattices = create_dimelement_sublattices(d)
+    vertexStates = map(x -> create_index_sublattice_state(standardBasis, x), allSubLattices)
 
     return vertexStates
 end
 
 """
-    createKernelHPolytope(vertexCoordinates)
+    create_kernel_hpolytope(vertexCoordinates)
 
 Return LazySets.HPolytope representation of polytope defined by `vertexCoordinates`.
 """
-function createKernelHPolytope(vertexCoordinates)
+function create_kernel_hpolytope(vertexCoordinates)
 
     #Create polytope in vertex representation
     vPt = VPolytope(vertexCoordinates)
@@ -28,25 +28,25 @@ end
 
 
 """
-    createKernelPolytope(d, standardBasis::StandardBasis)
+    create_kernel_polytope(d, standardBasis::StandardBasis)
 
 Return LazySets.HPolytope representation of the kernel polytope for dimension `d` and Bell basis `standardBasis`.
 """
-function createKernelPolytope(d, standardBasis::StandardBasis)
+function create_kernel_polytope(d, standardBasis::StandardBasis)
 
     # Create kernel vertex states 
-    vertexStates = createKernelVertexStates(d, standardBasis)
+    vertexStates = create_kernel_vertexstates(d, standardBasis)
 
     # Get coordinates 
     vertexCoords = map(x -> x[2], vertexStates)
 
     # Create kernel polytope 
-    return createKernelHPolytope(vertexCoords)
+    return create_kernel_hpolytope(vertexCoords)
 
 end
 
 """
-    extendSepVPolytopeBySepStates(
+    extend_vpolytope_by_densitystates(
         sepPolytope::VPolytope{Float64,Array{Float64,1}},
         sepDensityStates::Array{DensityState},
         precision::Integer
@@ -54,7 +54,7 @@ end
 
 Return an extended Lazysets.VPolytope representation of polytope of separable states based on given polytope `sepPolytope` and new separable `sepDensityStates` as new vertices.
 """
-function extendSepVPolytopeBySepStates(
+function extend_vpolytope_by_densitystates(
     sepPolytope::VPolytope{Float64,Array{Float64,1}},
     sepDensityStates::Array{DensityState},
     precision::Integer
@@ -62,11 +62,11 @@ function extendSepVPolytopeBySepStates(
 
     existingVertices = vertices_list(sepPolytope)
 
-    newVertices = unique(map(x -> roundDigits(abs.(x.coords), precision), sepDensityStates))
+    newVertices = unique(map(x -> rounddigits(abs.(x.coords), precision), sepDensityStates))
 
     combinedVertices = [existingVertices; newVertices]
 
-    uniqueVertices = unique(map(x -> roundDigits(abs.(x), precision), combinedVertices))
+    uniqueVertices = unique(map(x -> rounddigits(abs.(x), precision), combinedVertices))
 
     return VPolytope(uniqueVertices)
 
