### 2022-12-14 #4924 @meandmytram
Hey @kungfugo, the example shown in the documentation doesn't work for me. However, it works if I call `CreateStandardIndexbasis(d,n)` instead of `create_standard_indexbasis(d,n)`. Can you update the documentation on https://kungfugo.github.io/BellDiagonalQudits.jl/dev/manual/ please?
