### 2022-12-14 #4924 @meandmytram
Also, the first phrase of the paper's summary bugs me a little.

> In the field of quantum information and technology, entanglement of quantum states called qudits is regarded as resource for quantum and classical information processing tasks and allows the use of algorithms with better performance than any classical algorithm for certain applications like superdense coding, teleportation or computing.

1. Qudits are not quantum states but rather quantum systems.
2. "Entanglement as a resource for classical information processing task" sounds a bit weird I think.
3. I'd rather say "any known classical algorithm".
4. It's not really clear what a classical algorithm for quantum teleportation would be. In order to evade confusion, I would say something like "Entanglement of quantum states is often regarded as a computational resource for quantum computers opening possibilities for speedups in a variety of algorithms."
