### 2022-12-14 #4924 @kungfugo
> Hey @kungfugo, the example shown in the documentation doesn't work for me. However, it works if I call `CreateStandardIndexbasis(d,n)` instead of `create_standard_indexbasis(d,n)`. Can you update the documentation on https://kungfugo.github.io/BellDiagonalQudits.jl/dev/manual/ please?

Hey @meandmytram, have you pulled the latest changes? The source code only contains the function `create_standard_indexbasis`. `createStandardIndexBasis()` should not exist anymore.
