### 2022-12-15 #4924 @kungfugo
> > > Hey @kungfugo, the example shown in the documentation doesn't work for me. However, it works if I call `CreateStandardIndexbasis(d,n)` instead of `create_standard_indexbasis(d,n)`. Can you update the documentation on https://kungfugo.github.io/BellDiagonalQudits.jl/dev/manual/ please?
> >
> >
> > Hey @meandmytram, have you pulled the latest changes? The source code only contains the function `create_standard_indexbasis`. `createStandardIndexBasis()` should not exist anymore.
>
> I installed the package via `Pkg`, should I rather pull the package from GitHub?

This is my mistake. The latest version has not been tagged correctly. I will fix this asap. In the meanwhile you could pull from GitHub. There were no functional changes, only CI, tests and styling. Sorry for this problem.
