### 2022-12-15 #4924 @kungfugo
@meandmytram  I released a new version (v0.1.2)  compatible with the documentation and including your suggestions for the paper.
