### 2022-12-15 #4924 @meandmytram
@kungfugo Thanks for the changes, everything now seems to work according to documentation. One last but not the least thing I am a bit confused about is what it is exactly you're doing in the example. There is a lot of terminology I am not familiar with, sorry. From my understanding, you are building something like an ensemble of states from which you are then sampling and then analysing the entanglement, is this right? If that's something not too hard, could you please add a bit more explanations so that the content is accessible to a bit more general public.
