### 2023-01-02 d229782
Improved manual, new version v.0.1.3
