+++ docs/src/manual.md
@@ -1,5 +1,7 @@
 # Manual
 
+This manual shows how to use the package by using it to sample a set of uniformly distributed Bell diagonal states and to analyse their entanglement properties. Here, we apply criteria for separability and entanglement to determine the entanglement class of generated bipartite qutrits, i.e. `d=3`. In this system, bound entangled states, i.e. entangled states with positive partial transposition (PPT) that cannot be used for entanglement distillation, exist. The entanglement classes are labeled "SEP" for separability, "BOUND" for bound entanglement, "FREE" for entanglement with negative partial transposition (i.e. distillable) and "PPT_UNKNOWN" for PPT states that could not be classified as entangled or separable.
+
 ## Package installation
 
 BellDiagonalQudits can be installed using the Julia package manager. From the Julia REPL, type `]` to enter the Pkg REPL mode and run
@@ -16,14 +18,15 @@ julia> using BellDiagonalQudits
 
 ## State generation
 
-Create a basis of maximally entangled Bell states in `d` dimensions, indexed by the corresponding Weyl operator and sample some random states represented by `d^2` coordinates in the "enclosurePolytope", which is known to contain all PPT states.
+Create a basis of maximally entangled bipartite Bell states in `d^2` dimensions. Each Bell basis state is created by applying a certain Weyl transformation to the maximally entangled state. Sample Bell diagonal random mixed states of those Bell states, represented by their `d^2` coordinates (mixing probabilities) in the Bell basis.
 
 \
 **Bell basis generation**
 
-Create a basis and a dictionary to relate the `d^2` coordinates of a Bell diagonal state to the double indices `(k,l)` of the corresponding Weyl operator `W_{k,l}`.
+Create a Bell basis `myBasis` by applying each of the `d^2` Weyl transformations `W_{k,l} \\otimes \\mathbb{1}_d` to the maximally entangled state.`myBasis.basis` contains the enumerated Bell states in computational basis together with the indices of the corresponding Weyl transformation. `myBasisDict` contains the dictionaries to relate the enumerated `d^2` Bell basis states to the double indices `(k,l)` of the corresponding Weyl transformation.
 
 ```julia
+d = 3
 myBasis = create_standard_indexbasis(d,10)
 myBasisDict = create_dictionary_from_basis(myBasis)
 ```
@@ -31,13 +34,13 @@ myBasisDict = create_dictionary_from_basis(myBasis)
 \
 **State sampling**
 
-Create uniformly distributed random representations of quantum states by specifying the coordinates of the state in the created Bell basis.
+Create uniformly distributed random representations of quantum states by specifying the coordinates of the state in the created Bell basis. The coordinates represent the mixing probabilities of the Bell basis states. Here, we create only states with Bell coordinates within the "enclosure polytope", which is defined by the limitation of all coordinates (mixing probabilities) to be smaller than or equal to `1/d`. This subset is known to contain all (but not only) states with positive partial transposition (PPT), which can be separable or bound entangled.
 
 ```julia
 myCoordStates = uniform_bell_sampler(100, d, "enclosurePolytope")
 ```
 
-Create `DensityState`s including the density matrix represented in the computational basis.
+Create `DensityState`s including the density matrix of each state in the computational basis, created by mixing the Bell states of `myBasis` according to the `coords` of `myCoordStates`.
 
 ```julia
 myDensityStates = map(x->create_densitystate(x, myBasis), myCoordStates)
@@ -45,18 +48,18 @@ myDensityStates = map(x->create_densitystate(x, myBasis), myCoordStates)
 
 ## Analysis prerequisites
 
[…]
