+++ Project.toml
@@ -1,7 +1,7 @@
 name = "BellDiagonalQudits"
 uuid = "33287015-dcc4-41a7-aecb-0a84b862e0d8"
 authors = ["Christopher Popp <poppchristopher@pm.me>"]
-version = "0.1.2"
+version = "0.1.3"
 
 [deps]
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
