+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"library/#Library","page":"Library","title":"Library","text":"","category":"section"},{"location":"library/","page":"Library","title":"Library","text":"","category":"page"},{"location":"library/#Public","page":"Library","title":"Public","text":"","category":"section"},{"location":"library/","page":"Library","title":"Library","text":"Modules = [BellDiagonalQudits]\nPrivate = false","category":"page"},{"location":"library/#BellDiagonalQudits.AnalysedCoordState","page":"Library","title":"BellDiagonalQudits.AnalysedCoordState","text":"Representing an entanglement analyzed Bell diagonal state.\n\ncoordState: The analyzed CoordState\nkernel: true if kernel check successful, else false. missing if entanglement check not applied.\nspinrep: true if spinrep check successful, else false. missing if entanglement check not applied.\nppt: true if ppt check successful, else false. missing if entanglement check not applied.\nrealign: true if realignment check successful, else false. missing if entanglement check not applied.\nconcurrence: true if concurrence check successful, else false. missing if entanglement check not applied.\nmub: true if mub check successful, else false. missing if entanglement check not applied.\nnumericEW: true if numericEW check successful, else false. missing if entanglement check not applied.\n\n\n\n\n\n","category":"type"},{"location":"library/#BellDiagonalQudits.AnalysisSpecification","page":"Library","title":"BellDiagonalQudits.AnalysisSpecification","text":"Specification which entanglement checks to use.\n\nkernel_check\nspinrepCheck\nppt_check\nrealignment_check \nmub_check \nnumericewcheck\nuseSymmetries\n\n\n\n\n\n","category":"type"},{"location":"library/#BellDiagonalQudits.BoundedCoordEW","page":"Library","title":"BellDiagonalQudits.BoundedCoordEW","text":"Represents an entanglement witness W with extrema to detect entangled Bell diagonal states.\n\ncoords: Coordinates in Bell basis.\nupperBound: Upper bound of tr W rho satisfied by all separable states rho. Violation detects entanglement.\nlowerBound: Lower bound of tr W rho satisfied by all separable states rho. Violation detects entanglement.\ncheckedIterations: Number of iterations used in the optimization of bounds.\n\n\n\n\n\n","category":"type"},{"location":"library/#BellDiagonalQudits.ClassConflictException","page":"Library","title":"BellDiagonalQudits.ClassConflictException","text":"Exception for conflicts in analysis results.\n\nstate: The AnalysedCoordState for which a conflict occurs.\n\n\n\n\n\n","category":"type"},{"location":"library/#BellDiagonalQudits.CoordState","page":"Library","title":"BellDiagonalQudits.CoordState","text":"Represents a Bell diagonal state in Bell basis.\n\ncoords: Coordinates in Bell basis. \neClass: Entanglement class of the represented state.\n\n\n\n\n\n","category":"type"},{"location":"library/#BellDiagonalQudits.DensityState","page":"Library","title":"BellDiagonalQudits.DensityState","text":"Represents a Bell diagonal state.\n\ncoords: Coordinates in Bell basis.\ndensityMatrix: Hermitian density matrix in computational basis.\neClass: Entanglement class of the represented state.\n\n\n\n\n\n","category":"type"},{"location":"library/#BellDiagonalQudits.StandardBasis","page":"Library","title":"BellDiagonalQudits.StandardBasis","text":"Represents a Bell b
[…]
