### 2023-01-02 #4924 @kungfugo
@meandmytram , thanks for your comment. I agree.
I have released a new version (v.0.1.3) with a more detailed manual.
