### 2023-01-03 #4924 @Roger-luo
Hi @kungfugo I have some minor comments on the APIs and manual,

**mode selection** it might be better and more consistent with other Julia packages to just use `Symbol` for mode selection, e.g

```julia
myCoordStates = uniform_bell_sampler(100, d, "enclosurePolytope")
```

can be

```julia
myCoordStates = uniform_bell_sampler(100, d, :enclosurePolytope)
```

it is not a big difference tho, but you can see [this discussion](https://discourse.julialang.org/t/when-should-a-function-accept-a-symbol-as-an-argument/43510/17?u=roger-luo) to understand why

**`tovrep` is not (re)exported**, this causing manual example to fail. Tho, this is a method from `LazySets` but I believe you should re-export this function, otherwise, you need to show your user they need to `using LazySets`. Please consider putting a [doc test](https://documenter.juliadocs.org/stable/man/doctests/) in your tests to prevent an example from failing again.

I think it is better to make the following example in the manual a self-contained example, the current one seems a bit hand-waving - one cannot directly run that code block without defining
a few things extra

```julia
myExtendedKernel = extend_vpolytope_by_densitystates(tovrep(mySepKernel), newSepDensityStates)
```

A similar issue exists for

```julia
myOptimizedEWs = create_random_bounded_ews(
    d,
    myBasis,
    n,
    true,
    50
    )
```

I'd suggest you to go through these examples and making sure clicking copy-paste buttons gives runnable scripts instead of just a code block.

In summary, I think the functionality is solid, it does what is described.
The documentation requires further improvement as commented above.

And a non-blocking suggestion from a user perspective, the current APIs are too long, and some of them seems to have too much positional arguments, e.g

```julia
f(x) = analyse_coordstate(
    d,
    x,
    myAnaSpec,
    myBasis,
    mySepKernel,
    myWeylOperatorBasis,
    myBasisDict,
    missing,
    myOptimizedCoodEWs
)
```

or

```julia
myAnaSpec = AnalysisSpecification(
   true,
   true,
   true,
   true,
   true,
   false,
   true,
   false
)
```

If you need to specify configuration/options, I'd suggest checking out https://github.com/Roger-luo/Configurations.jl which is designed for the case. But if you are able to make it more automatic, or incremental (instead of putting a bunch of `true` all in one) it will have a much friendlier UX.
