### 2023-01-04 #4924 @meandmytram
The example looks way less opaque now, I think once we're done with Roger's comments we'd be all set.
