### 2023-01-04 aeb73a9
New version 0.1.4, mode selection via symbols, manual improvements
