### 2023-01-04 #4924 @kungfugo
Hi @Roger-luo,
thank you for your comments. I released v0.1.4 with changes according to your comments. In particular:

- Mode selection: Symbols are now used instead of strings.
- Exported functions: `tovrep` is now exported as well
- Manual: The manual is now self-contained. All code can now be run by copy-and-paste. minor improvements

Concerning your comments regarding `doctest`:
I understand the issue, however, the output of the manual is not deterministic due to the random sampling. As far as I understand, this prevents the straightforward use of `doctest`. For this release, I would prefer to rely on the tested current implementation, but I will consider this issue for the next major release.

Concerning your comments regarding the API:
I agree that there can be significant improvements concerning the UX. However, for this release, I prefer to keep the current implementation for the sake of flexibility of use and stability of code, but I will consider your suggestions for the next major release. The same goes for `Configurations.jl`, which seems very suitable to improve the UX.
