### 2023-01-04 #4924 @Roger-luo
> I understand the issue, however, the output of the manual is not deterministic due to the random sampling.

You can set up seed at the beginning of doctest, just copy-paste the following to your `runtests.jl`

```julia
using Documenter
DocMeta.setdocmeta!(BellDiagonalQudits, :DocTestSetup, :(using BellDiagonalQudits, Random;Random.seed!(123)); recursive=true)
doctest(BellDiagonalQudits)
```

This reminds me that you should have an explicit argument for RNG for methods calling `rand` etc, e.g this line https://github.com/kungfugo/BellDiagonalQudits.jl/blob/main/src/BellStates/bellStates.jl#L16
should have an argument `rng::AbstractRNG` otherwise one will not be able to set RNG state explicitly thus issues like the above in a doctest will appear.

But you can do it later, while this package still seems to require a lot of improvements. I think the minimum requirements are met now.
