### 2023-01-27 #4924 @kungfugo
> Great! Do you want to make any further edits based on what Roger is suggesting above?

I will incorporate those suggestions in the next major release.
For now I'd like to publish with JOSS based on the current version.
