### 2023-01-27 #4924 @jarvist
OK! I made a minor suggestion for the paper.

Have you generated a DOI for the code yet? Probably easiest would be to use Zenodo to publish a tagged version of your code.
>Upon successful completion of the review, authors will make a tagged release of the software, and deposit a copy of the repository with a data-archiving service such as [Zenodo](https://zenodo.org/) or [figshare](https://figshare.com/), get a DOI for the archive, and update the review issue thread with the version number and DOI.
