### 2023-01-27 #4924 @jarvist
> Sorry, but what suggestion do you mean?

https://github.com/kungfugo/BellDiagonalQudits.jl/compare/main...jarvist:patch-1?expand=1
