### 2023-01-27 #4924 @kungfugo
Allright!
Included your suggestions, created release v0.1.5 and registered with Zenodo,
[](https://doi.org/10.5281/zenodo.7575767)
