### 2023-01-27 #4924 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
