### 2023-01-27 #4924 @kungfugo
@jarvist , @Roger-luo , @meandmytram  thank you for your efforts! First time publishing with JOSS and it was a nice ecperience. Looking forward to the next ones.
