### 2023-01-29 #4924 @kyleniemeyer
Hi @kungfugo, just doing some final checks before accepting.

- [ ] Could you clean up the Zenodo metadata? The author list does not have your real name, and @jarvist is listed as a contributor—generally this should match the paper.
- [ ] Can you remove the "These authors contributed equally." bit from the author block of the paper? It doesn't really make sense for a single-author paper.
- [ ] In a few places, you use a reference as part of a sentence, but with the parenthetical format. Could you replace this with the "Author et al. (year)" format? For example, rather than "summarized in (Baumgartner et al., 2007)", replace with "summarized by Baumgartner et al. (2007)". This is done with the `@author:year` syntax.
