### 2023-01-30 66ca33b
Add Zenodo metadata, paper citations, version number
