### 2023-01-30 e2c397b
Paper citation styling
