### 2023-01-30 #4924 @kungfugo
Hi @kyleniemeyer ,

thank you for your comments. I have a released a new version v0.1.6 including your requests and registered it with zenodo.

In particular:

- New DOI with corrected Zenodo metadata: DOI: [10.5281/zenodo.7585987 ](10.5281/zenodo.7585987)
- The author block does not the 'equal contribution' part anymore
- References cited without parentheses when part of a sentence
