### 2023-01-31 #4924 @arfon
> (I'm going to reopen because it looks like the PDF is not yet appearing—I saw there were some systemic issues with GitHub Actions earlier.)

Should be fixed now!
